agent/pyproject.toml
[project]
name = "order-to-cash"
version = "0.1.0"
description = "SAP Order to Cash automation with Google ADK multi-agent system."
authors = [
    { name = "Felipe Lujan" },
]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.16.1",
    "python-dotenv>=1.0.1",
    "litellm>=1.0.0",
    "anthropic>=0.40.0",
]

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "google-adk[eval]>=1.0.0",
    "pytest-asyncio>=0.26.0",
]
